Move `DatabaseContainerMock` to separate file
